snsf.oa.ror.zurich.26: Code and data for an analysis of the SNSF OA monitoring presented at ROR Zürich 2026

This package contains the code and data to reproduce an analysis presented at the Research on Research Event in Zürich on 2026/01/29. The analysis focuses on the coverage, within different academic databases, of scientific publications identified from Swiss National Science Foundation (SNSF) grants. Identifying these publications is critical for monitoring the open access of publications from SNSF grants.

See also

Author

Maintainer: Simon Gorin simon.gorin@snf.ch (ORCID)